IoT-243 Connectivity Abstraction single-threaded module missing definition for CASendRequestToAll().
IoT-161 PCF Command is not used in OIC_LOG() calls under the connectivity directory.

Updated documentation and few memory leaks .

Reviewed-on: https://gerrit.iotivity.org/gerrit/142
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>

+++ resource/csdk/connectivity/src/ethernet_adapter/arduino/caethernetclient.cpp
@@ -60,7 +60,7 @@ void CAEthernetSetUnicastSocket(const int32_t socketID)
     return;
 }
 
-void CAEthernetSetUnicastPort(const int32_t port)
+void CAEthernetSetUnicastPort(const int16_t port)
 {
     OIC_LOG(DEBUG, MOD_NAME, "IN");
     if (0 < port)
